requirements.txt
altgraph==0.17
astroid==2.3.3
fbs==0.8.6
future==0.18.2
isort==4.3.21
joblib==0.11
lazy-object-proxy==1.4.3
llvmlite==0.31.0
lxml
macholib==1.14
mccabe==0.6.1
numpy
pefile==2019.4.18
Pillow==7.0.0
pylint==2.4.4
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
six==1.14.0
typed-ast==1.4.1
wrapt==1.11.2
